continue improving the examples
