+++ docs/src/pages/guides/routing/routing.md
@@ -97,7 +97,7 @@ const LinkBehavior = React.forwardRef((props, ref) => (
 Next.js has [a custom Link component](https://nextjs.org/docs/api-reference/next/link).
 The [example folder](https://github.com/mui-org/material-ui/tree/HEAD/examples/nextjs-with-typescript) provides adapters for usage with MUI.
 
-- The first version of the adapter is the [`NextLinkComposed`](https://github.com/mui-org/material-ui/blob/HEAD/examples/nextjs-with-typescript/src/components/Link.tsx) component.
+- The first version of the adapter is the [`NextLinkComposed`](https://github.com/mui-org/material-ui/blob/HEAD/examples/nextjs-with-typescript/src/Link.tsx) component.
   This component is unstyled and only responsible for handling the navigation.
   The prop `href` was renamed `to` to avoid a naming conflict.
   This is similar to react-router's Link component.
