--- examples/nextjs-with-typescript/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "env": {
-    "browser": true
-  }
-}
