+++ examples/nextjs-with-typescript/pages/about.tsx
@@ -3,11 +3,12 @@ import type { NextPage } from 'next';
 import Container from '@mui/material/Container';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
-import Link from '../components/Link';
-import ProTip from '../components/ProTip';
-import Copyright from '../components/Copyright';
+import Button from '@mui/material/Button';
+import Link from '../src/Link';
+import ProTip from '../src/ProTip';
+import Copyright from '../src/Copyright';
 
-const Home: NextPage = () => {
+const About: NextPage = () => {
   return (
     <Container maxWidth="lg">
       <Box
@@ -22,9 +23,11 @@ const Home: NextPage = () => {
         <Typography variant="h4" component="h1" gutterBottom>
           MUI v5 + Next.js with TypeScript example
         </Typography>
-        <Link href="/about" color="secondary">
-          Go to the about page
-        </Link>
+        <Box maxWidth="sm">
+          <Button variant="contained" component={Link} noLinkStyle href="/">
+            Go to the home page
+          </Button>
+        </Box>
         <ProTip />
         <Copyright />
       </Box>
@@ -32,4 +35,4 @@ const Home: NextPage = () => {
   );
 };
 
-export default Home;
+export default About;
