+++ examples/nextjs-with-typescript/pages/index.tsx
@@ -3,12 +3,11 @@ import type { NextPage } from 'next';
 import Container from '@mui/material/Container';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import Link from '../components/Link';
-import ProTip from '../components/ProTip';
-import Copyright from '../components/Copyright';
+import Link from '../src/Link';
+import ProTip from '../src/ProTip';
+import Copyright from '../src/Copyright';
 
-const About: NextPage = () => {
+const Home: NextPage = () => {
   return (
     <Container maxWidth="lg">
       <Box
@@ -23,11 +22,9 @@ const About: NextPage = () => {
         <Typography variant="h4" component="h1" gutterBottom>
           MUI v5 + Next.js with TypeScript example
         </Typography>
-        <Box maxWidth="sm">
-          <Button variant="contained" component={Link} noLinkStyle href="/">
-            Go to the home page
-          </Button>
-        </Box>
+        <Link href="/about" color="secondary">
+          Go to the about page
+        </Link>
         <ProTip />
         <Copyright />
       </Box>
@@ -35,4 +32,4 @@ const About: NextPage = () => {
   );
 };
 
-export default About;
+export default Home;
