--- examples/nextjs-with-typescript/sandbox.config.json
@@ -1,5 +1,0 @@
-{
-  "container": {
-    "node": "16"
-  }
-}
