+++ examples/nextjs-with-typescript/src/Copyright.tsx
