+++ examples/nextjs-with-typescript/src/ProTip.tsx
