--- examples/nextjs/sandbox.config.json
@@ -1,5 +1,0 @@
-{
-  "container": {
-    "node": "12"
-  }
-}
