+++ examples/remix-with-typescript/.stackblitzrc
@@ -1,0 +1,3 @@
+{
+  "startCommand": "npm run dev"
+}
